requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-sendsms==0.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.7.7
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
setuptools-rust==0.11.6
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
